'author' property added in recipe

+++ foodapp/routes.py
@@ -6,33 +6,39 @@ from foodapp.models import User, Recipe, Ingredient
 from foodapp import app, bcrypt, db
 from foodapp.forms import RegistrationForm, LoginForm, AddRecipe
 from foodapp.utils import searching_by_dish_name, filter, findRecipe
-from flask_login import login_user, logout_user
+from flask_login import login_user, logout_user, current_user
 
 
 @app.route("/", methods=['GET', 'POST'])
 def reg():
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        hash_pwd = bcrypt.generate_password_hash(form.password.data).decode('utf-8')
-        usr = User(username=form.username.data, email=form.email.data, password=hash_pwd)
-        db.session.add(usr)
-        db.session.commit()
-        flash('Your account has been created. Please login to continue', 'success')
-        return redirect(url_for('login'))
-    return render_template('register.html', form=form)
+    if current_user.is_authenticated :
+        return redirect(url_for('home'))
+    else :
+        form = RegistrationForm()
+        if form.validate_on_submit():
+            hash_pwd = bcrypt.generate_password_hash(form.password.data).decode('utf-8')
+            usr = User(username=form.username.data, email=form.email.data, password=hash_pwd)
+            db.session.add(usr)
+            db.session.commit()
+            flash('Your account has been created. Please login to continue', 'success')
+            return redirect(url_for('login'))
+        return render_template('register.html', form=form)
 
 @app.route("/login", methods=['POST', 'GET'])
 def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data).first()
-        if user and bcrypt.check_password_hash(user.password, form.password.data) :
-            login_user(user, remember=form.remember.data)
-            flash('Login successful, Welcome ' + user.username +' !', 'success')
-            return redirect(url_for('home'))
-        else :
-            flash('The email or password you have entered is incorrect. Please try again', 'danger')
-    return render_template('login.html', form=form)
+    if current_user.is_authenticated :
+        return redirect(url_for('home'))
+    else :
+        form = LoginForm()
+        if form.validate_on_submit():
+            user = User.query.filter_by(email=form.email.data).first()
+            if user and bcrypt.check_password_hash(user.password, form.password.data) :
+                login_user(user, remember=form.remember.data)
+                flash('Login successful, Welcome ' + user.username +' !', 'success')
+                return redirect(url_for('home'))
+            else :
+                flash('The email or password you have entered is incorrect. Please try again', 'danger')
+        return render_template('login.html', form=form)
 
 @app.route('/logout')
 def logout():
@@ -42,8 +48,13 @@ def logout():
 
 @app.route("/home", methods=['GET', 'POST'])
 def home():
-    recipes = Recipe.query.all()
-    return render_template('home.html', recipes=recipes)
+    if current_user.is_authenticated :
+        recipes = Recipe.query.all()
+        flash('Login successful, Welcome ' + User.query.get(current_user.get_id()).username +' !', 'success')
+        return render_template('home.html', recipes=recipes)
+    else :
+        flash('Please login', 'info')
+        return redirect(url_for('login'))
 
 def save_picture(form_picture) :
     ran_hex = secrets.token_hex(8)
@@ -60,44 +71,50 @@ def save_picture(form_picture) :
 
 @app.route("/add", methods=['GET', 'POST'])
 def add():
-    form = AddRecipe()
-    theCheckList = []
-    if request.method == 'POST':
-        theCheckList = request.form.getlist('ingredient')
-        print(theCheckList)
-    if form.validate_on_submit():
-        recipe = Recipe(name=form.recipe_name.data, procedure=form.procedure.data)
-        if form.picture.data :
-            picture_file = save_picture(form.picture.data)
-            recipe.image_file = picture_file
-        db.session.add(recipe)
-        db.session.commit()
-        
-        for i in theCheckList:
-            if Ingredient.query.filter_by(name=i).count() == 0 :
-                print("fdsaf")
-                new_ing = Ingredient(name=i)
-                db.session.add(new_ing)
-                db.session.commit()
-                recipe.ingredients.append(new_ing)
-                db.session.commit()
+    if current_user.is_authenticated :
+        form = AddRecipe()
+        theCheckList = []
+        if request.method == 'POST':
+            theCheckList = request.form.getlist('ingredient')
+            print(theCheckList)
+        if form.validate_on_submit():
+            recipe = Recipe(name=form.recipe_name.data, procedure=form.procedure.data)
+            if form.picture.data :
+                picture_file = save_picture(form.picture.data)
+                recipe.image_file = picture_file
+            db.session.add(recipe)
+            db.session.commit()
+            recipe.author = User.query.get(current_user.get_id())
+            
+            for i in theCheckList:
+                if Ingredient.query.filter_by(name=i).count() == 0 :
+                    print("fdsaf")
+                    new_ing = Ingredient(name=i)
+                    db.session.add(new_ing)
+                    db.session.commit()
+                    recipe.ingredients.append(new_ing)
+                    db.session.commit()
 
-            else :
-                recipe.ingredients.append(Ingredient.query.filter_by(name=i).first())
-                db.session.commit()
-                
+                else :
+                    recipe.ingredients.append(Ingredient.query.filter_by(name=i).first())
+                    db.session.commit()
+                    
 
-        return redirect(url_for('add'))
-
-    return render_template('add.html', form=form)
+            return redirect(url_for('add'))
+        flash('Please login', 'info')
+        return render_template('add.html', form=form)
+    else :
+        return redirect(url_for('login'))
 
 @app.route("/delete/<id>", methods=['GET', 'POST'])
 def delete(id):
     recipe = Recipe.query.get(id)
-    db.session.delete(recipe)
-    db.session.commit()
-
-    return jsonify({ 'data' : 'none' })
+    if recipe.author == User.query.get(current_user.get_id()) :
+        db.session.delete(recipe)
+        db.session.commit()
+        return jsonify({ 'data' : 'none' })
+    else :
+        return "Unauthorized"
 
 @app.route('/ingredientsearch', methods=['POST', 'GET'])
 def ingredientsearch() :
@@ -121,8 +138,11 @@ def search_by_dish():
 
 @app.route('/recipe', methods=['POST', 'GET'])
 def recipeFiltering() :
-    recipes = Recipe.query.all()
-    return render_template('recipe.html')
+    if current_user.is_authenticated :
+        recipes = Recipe.query.all()
+        return render_template('recipe.html')
+    else :
+        return redirect(url_for('login'))
 
 @app.route('/filter', methods=['POST', 'GET'])
 def filteredrecipe() :
